fix: handle NodeClaim termination in the lifecycle controller
