+++ pkg/controllers/nodeclaim/lifecycle/controller.go
@@ -18,32 +18,38 @@ package lifecycle
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	"github.com/patrickmn/go-cache"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/samber/lo"
 	"go.uber.org/multierr"
 	"golang.org/x/time/rate"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/client-go/util/workqueue"
+	"k8s.io/klog/v2"
 	"k8s.io/utils/clock"
 	controllerruntime "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/event"
+	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+
+	terminatorevents "sigs.k8s.io/karpenter/pkg/controllers/node/termination/terminator/events"
 
 	v1 "sigs.k8s.io/karpenter/pkg/apis/v1"
 	"sigs.k8s.io/karpenter/pkg/cloudprovider"
 	"sigs.k8s.io/karpenter/pkg/events"
+	"sigs.k8s.io/karpenter/pkg/metrics"
 	"sigs.k8s.io/karpenter/pkg/operator/injection"
 	nodeclaimutil "sigs.k8s.io/karpenter/pkg/utils/nodeclaim"
 	"sigs.k8s.io/karpenter/pkg/utils/result"
+	terminationutil "sigs.k8s.io/karpenter/pkg/utils/termination"
 )
 
 type nodeClaimReconciler interface {
@@ -55,7 +61,9 @@ type nodeClaimReconciler interface {
 // the cluster as nodes and that they are properly initialized, ensuring that nodeclaims that do not have matching nodes
 // after some liveness TTL are removed
 type Controller struct {
-	kubeClient client.Client
+	kubeClient    client.Client
+	cloudProvider cloudprovider.CloudProvider
+	recorder      events.Recorder
 
 	launch         *Launch
 	registration   *Registration
@@ -65,7 +73,9 @@ type Controller struct {
 
 func NewController(clk clock.Clock, kubeClient client.Client, cloudProvider cloudprovider.CloudProvider, recorder events.Recorder) *Controller {
 	return &Controller{
-		kubeClient: kubeClient,
+		kubeClient:    kubeClient,
+		cloudProvider: cloudProvider,
+		recorder:      recorder,
 
 		launch:         &Launch{kubeClient: kubeClient, cloudProvider: cloudProvider, cache: cache.New(time.Minute, time.Second*10), recorder: recorder},
 		registration:   &Registration{kubeClient: kubeClient},
@@ -74,11 +84,35 @@ func NewController(clk clock.Clock, kubeClient client.Client, cloudProvider clou
 	}
 }
 
+func (c *Controller) Register(_ context.Context, m manager.Manager) error {
+	return controllerruntime.NewControllerManagedBy(m).
+		Named(c.Name()).
+		For(&v1.NodeClaim{}).
+		Watches(
+			&corev1.Node{},
+			nodeclaimutil.NodeEventHandler(c.kubeClient),
+		).
+		WithOptions(controller.Options{
+			RateLimiter: workqueue.NewTypedMaxOfRateLimiter[reconcile.Request](
+				// back off until last attempt occurs ~90 seconds before nodeclaim expiration
+				workqueue.NewTypedItemExponentialFailureRateLimiter[reconcile.Request](time.Second, 300*time.Second),
+				// 10 qps, 100 bucket size
+				&workqueue.TypedBucketRateLimiter[reconcile.Request]{Limiter: rate.NewLimiter(rate.Limit(10), 100)},
+			),
+			MaxConcurrentReconciles: 1000, // higher concurrency limit since we want fast reaction to node syncing and launch
+		}).
+		Complete(reconcile.AsReconciler(m.GetClient(), c))
+}
+
+func (c *Controller) Name() string {
+	return "nodeclaim.lifecycle"
+}
+
 func (c *Controller) Reconcile(ctx context.Context, nodeClaim *v1.NodeClaim) (reconcile.Result, error) {
-	ctx = injection.WithControllerName(ctx, "nodeclaim.lifecycle")
+	ctx = injection.WithControllerName(ctx, c.Name())
 
 	if !nodeClaim.DeletionTimestamp.IsZero() {
-		return reconcile.Result{}, nil
+		return c.finalize(ctx, nodeClaim)
 	}
 
 	// Add the finalizer immediately since we shouldn't launch if we don't yet have the finalizer.
@@ -130,28 +164,108 @@ func (c *Controller) Reconcile(ctx context.Context, nodeClaim *v1.NodeClaim) (re
 	return result.Min(results...), nil
 }
 
-func (c *Controller) Register(_ context.Context, m manager.Manager) error {
-	return controllerruntime.NewControllerManagedBy(m).
-		Named("nodeclaim.lifecycle").
-		For(&v1.NodeClaim{}, builder.WithPredicates(
-			predicate.Funcs{
-				CreateFunc: func(e event.CreateEvent) bool { return true },
-				UpdateFunc: func(e event.UpdateEvent) bool { return false },
-				DeleteFunc: func(e event.DeleteEvent) bool { return false },
-			},
-		)).
-		Watches(
-			&corev1.Node{},
-			nodeclaimutil.NodeEventHandler(c.kubeClient),
-		).
-		WithOptions(controller.Options{
-			RateLimiter: workqueue.NewTypedMaxOfRateLimiter[reconcile.Request](
-				// back off until last attempt occurs ~90 seconds before nodeclaim expiration
-				workqueue.NewTypedItemExponentialFailureRateLimiter[reconcile.Request](time.Second, 300*time.Second),
-				// 10 qps, 100 bucket size
-				&workqueue.TypedBucketRateLimiter[reconcile.Request]{Limiter: rate.NewLimiter(rate.Limit(10), 100)},
-			),
-			MaxConcurrentReconciles: 1000, // higher concurrency limit since we want fast reaction to node syncing and launch
-		}).
-		Complete(reconcile.AsReconciler(m.GetClient(), c))
+//nolint:gocyclo
+func (c *Controller) finalize(ctx context.Context, nodeClaim *v1.NodeClaim) (reconcile.Result, error) {
+	ctx = log.IntoContext(ctx, log.FromContext(ctx).WithValues("Node", klog.KRef("", nodeClaim.Status.NodeName), "provider-id", nodeClaim.Status.ProviderID))
+	if !controllerutil.ContainsFinalizer(nodeClaim, v1.TerminationFinalizer) {
+		return reconcile.Result{}, nil
+	}
+	if err := c.ensureTerminationGracePeriodTerminationTimeAnnotation(ctx, nodeClaim); err != nil {
+		return reconcile.Result{}, fmt.Errorf("adding nodeclaim terminationGracePeriod annotation, %w", err)
+	}
+
+	nodes, err := nodeclaimutil.AllNodesForNodeClaim(ctx, c.kubeClient, nodeClaim)
+	if err != nil {
+		return reconcile.Result{}, err
+	}
+	for _, node := range nodes {
+		// If we still get the Node, but it's already marked as terminating, we don't need to call Delete again
+		if !node.DeletionTimestamp.IsZero() {
+			continue
+		}
+		// We delete nodes to trigger the node finalization and deletion flow
+		if err = c.kubeClient.Delete(ctx, node); client.IgnoreNotFound(err) != nil {
+			return reconcile.Result{}, err
+		}
+	}
+	// We wait until all the nodes associated with this nodeClaim have completed their deletion before triggering the finalization of the nodeClaim
+	if len(nodes) > 0 {
+		return reconcile.Result{}, nil
+	}
+	var isInstanceTerminated bool
+	// We can expect ProviderID to be empty when there is a failure while launching the nodeClaim
+	if nodeClaim.Status.ProviderID != "" {
+		isInstanceTerminated, err = terminationutil.EnsureTerminated(ctx, c.kubeClient, nodeClaim, c.cloudProvider)
+		if err != nil {
+			// 404 = the nodeClaim no longer exists
+			if errors.IsNotFound(err) {
+				return reconcile.Result{}, nil
+			}
+			// 409 - The nodeClaim exists, but its status has already been modified
+			if errors.IsConflict(err) {
+				return reconcile.Result{Requeue: true}, nil
+			}
+			return reconcile.Result{}, fmt.Errorf("ensuring instance termination, %w", err)
+		}
+		if !isInstanceTerminated {
+			return reconcile.Result{RequeueAfter: 5 * time.Second}, nil
+		}
+		InstanceTerminationDurationSeconds.With(map[string]string{
+			metrics.NodePoolLabel: nodeClaim.Labels[v1.NodePoolLabelKey],
+		}).Observe(time.Since(nodeClaim.StatusConditions().Get(v1.ConditionTypeInstanceTerminating).LastTransitionTime.Time).Seconds())
+	}
+	stored := nodeClaim.DeepCopy() // The NodeClaim may have been modified in the EnsureTerminated function
+	controllerutil.RemoveFinalizer(nodeClaim, v1.TerminationFinalizer)
+	if !equality.Semantic.DeepEqual(stored, nodeClaim) {
+		// We use client.MergeFromWithOptimisticLock because patching a list with a JSON merge patch
+		// can cause races due to the fact that it fully replaces the list on a change
+		// Here, we are updating the finalizer list
+		// https://github.com/kubernetes/kubernetes/issues/111643#issuecomment-2016489732
+		if err = c.kubeClient.Patch(ctx, nodeClaim, client.MergeFromWithOptions(stored, client.MergeFromWithOptimisticLock{})); err != nil {
+			if errors.IsConflict(err) {
+				return reconcile.Result{Requeue: true}, nil
+			}
+			return reconcile.Result{}, client.IgnoreNotFound(fmt.Errorf("removing termination finalizer, %w", err))
+		}
+		log.FromContext(ctx).Info("deleted nodeclaim")
+		NodeClaimTerminationDurationSeconds.With(map[string]string{
+			metrics.NodePoolLabel: nodeClaim.Labels[v1.NodePoolLabelKey],
+		}).Observe(time.Since(stored.DeletionTimestamp.Time).Seconds())
+		metrics.NodeClaimsTerminatedTotal.With(prometheus.Labels{
+			metrics.NodePoolLabel:     nodeClaim.Labels[v1.NodePoolLabelKey],
+			metrics.CapacityTypeLabel: nodeClaim.Labels[v1.CapacityTypeLabelKey],
+		}).Inc()
+	}
+	return reconcile.Result{}, nil
+
+}
+
+func (c *Controller) ensureTerminationGracePeriodTerminationTimeAnnotation(ctx context.Context, nodeClaim *v1.NodeClaim) error {
+	// if the expiration annotation is already set, we don't need to do anything
+	if _, exists := nodeClaim.ObjectMeta.Annotations[v1.NodeClaimTerminationTimestampAnnotationKey]; exists {
+		return nil
+	}
+
+	// In Kubernetes, every object has a terminationGracePeriodSeconds, defaulted to and un-changeable from 0. There is an additional TerminationGracePeriodSeconds in the PodSpec which can be configured.
+	// We use the kubernetes object TerminationGracePeriod to infer that the DeletionTimestamp is always equal to the time the NodeClaim is deleted.
+	// This should not be confused with the NodeClaim.spec.terminationGracePeriod field introduced in Karpenter Custom Resources.
+	if nodeClaim.Spec.TerminationGracePeriod != nil && !nodeClaim.ObjectMeta.DeletionTimestamp.IsZero() {
+		terminationTimeString := nodeClaim.DeletionTimestamp.Time.Add(nodeClaim.Spec.TerminationGracePeriod.Duration).Format(time.RFC3339)
+		return c.annotateTerminationGracePeriodTerminationTime(ctx, nodeClaim, terminationTimeString)
+	}
+
+	return nil
+}
+
+func (c *Controller) annotateTerminationGracePeriodTerminationTime(ctx context.Context, nodeClaim *v1.NodeClaim, terminationTime string) error {
+	stored := nodeClaim.DeepCopy()
+	nodeClaim.ObjectMeta.Annotations = lo.Assign(nodeClaim.ObjectMeta.Annotations, map[string]string{v1.NodeClaimTerminationTimestampAnnotationKey: terminationTime})
+
+	if err := c.kubeClient.Patch(ctx, nodeClaim, client.MergeFrom(stored)); err != nil {
+		return client.IgnoreNotFound(err)
+	}
+	log.FromContext(ctx).WithValues(v1.NodeClaimTerminationTimestampAnnotationKey, terminationTime).Info("annotated nodeclaim")
+	c.recorder.Publish(terminatorevents.NodeClaimTerminationGracePeriodExpiring(nodeClaim, terminationTime))
+
+	return nil
 }
