+++ pkg/controllers/nodeclaim/lifecycle/metrics.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package termination
+package lifecycle
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
